+++ distros/base/.bash-config/functions/install/manual/generic.sh
@@ -1,20 +1,4 @@
 # !/bin/bash
- 
-# ==============================================================================
-# FUNÇÕES DE INSTALAÇÃO DE ...
-# ==============================================================================
-
-# Navegador: Google Chrome
-install_browser() {
-  echo -e "${b_cyan}[INFO]${nc} Instalando Google Chrome..."
-  
-  wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb || return 1
-  sudo apt install ./google-chrome*.deb -y
-  rm google-chrome*.deb
-  
-  echo -e "${b_green}[SUCESSO]${nc} Google Chrome instalado!"
-}
-
 
 # =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
 # Extração manual de alguns aplicativos (binários)